+++ src/env_posix.c
@@ -52,7 +52,7 @@ int mongo_env_write_socket( mongo *conn, const void *buf, int len ) {
 #endif
 
     while ( len ) {
-        int sent = send( conn->sock, cbuf, len, flags );
+        int sent = (int)send( conn->sock, cbuf, len, flags );
         if ( sent == -1 ) {
             if (errno == EPIPE)
                 conn->connected = 0;
@@ -69,7 +69,7 @@ int mongo_env_write_socket( mongo *conn, const void *buf, int len ) {
 int mongo_env_read_socket( mongo *conn, void *buf, int len ) {
     char *cbuf = buf;
     while ( len ) {
-        int sent = recv( conn->sock, cbuf, len, 0 );
+        int sent = (int)recv( conn->sock, cbuf, len, 0 );
         if ( sent == 0 || sent == -1 ) {
             __mongo_set_error( conn, MONGO_IO_ERROR, strerror( errno ), errno );
             return MONGO_ERROR;
