+++ src/mongo.c
@@ -121,7 +121,7 @@ MONGO_EXPORT void __mongo_set_error( mongo *conn, mongo_error_t err, const char 
     conn->errcode = errcode;
 
     if( str ) {
-        str_size = strlen( str ) + 1;
+        str_size = (int)strlen( str ) + 1;
         errstr_size = str_size > MONGO_ERR_LEN ? MONGO_ERR_LEN : str_size;
         memcpy( conn->errstr, str, errstr_size );
         conn->errstr[errstr_size-1] = '\0';
@@ -442,7 +442,7 @@ MONGO_EXPORT void mongo_replica_set_init( mongo *conn, const char *name ) {
     conn->replica_set->primary_connected = 0;
     conn->replica_set->seeds = NULL;
     conn->replica_set->hosts = NULL;
-    conn->replica_set->name = ( char * )bson_malloc( strlen( name ) + 1 );
+    conn->replica_set->name = ( char * )bson_malloc( (int)strlen( name ) + 1 );
     memcpy( conn->replica_set->name, name, strlen( name ) + 1  );
 
     conn->primary = bson_malloc( sizeof( mongo_host_port ) );
@@ -866,13 +866,13 @@ MONGO_EXPORT int mongo_insert( mongo *conn, const char *ns,
 
     mm = mongo_message_create( 16 /* header */
                                + 4 /* ZERO */
-                               + strlen( ns )
+                               + (int)strlen( ns )
                                + 1 + bson_size( bson )
                                , 0, 0, MONGO_OP_INSERT );
 
     data = &mm->data;
     data = mongo_data_append32( data, &ZERO );
-    data = mongo_data_append( data, ns, strlen( ns ) + 1 );
+    data = mongo_data_append( data, ns, (int)strlen( ns ) + 1 );
     mongo_data_append( data, bson->data, bson_size( bson ) );
 
 
@@ -898,7 +898,7 @@ MONGO_EXPORT int mongo_insert_batch( mongo *conn, const char *ns,
     mongo_write_concern *write_concern = NULL;
     int i;
     char *data;
-    int overhead =  16 + 4 + strlen( ns ) + 1;
+    int overhead =  16 + 4 + (int)strlen( ns ) + 1;
     int size = overhead;
 
     if( mongo_validate_ns( conn, ns ) != MONGO_OK )
@@ -927,7 +927,7 @@ MONGO_EXPORT int mongo_insert_batch( mongo *conn, const char *ns,
         data = mongo_data_append32( data, &ONE );
     else
         data = mongo_data_append32( data, &ZERO );
-    data = mongo_data_append( data, ns, strlen( ns ) + 1 );
+    data = mongo_data_append( data, ns, (int)strlen( ns ) + 1 );
 
     for( i=0; i<count; i++ ) {
         data = mongo_data_append( data, bsons[i]->data, bson_size( bsons[i] ) );
@@ -968,7 +968,7 @@ MONGO_EXPORT int mongo_update( mongo *conn, const char *ns, const bson *cond,
 
     mm = mongo_message_create( 16 /* header */
                                + 4  /* ZERO */
-                               + strlen( ns ) + 1
+                               + (int)strlen( ns ) + 1
                                + 4  /* flags */
                                + bson_size( cond )
                                + bson_size( op )
@@ -976,7 +976,7 @@ MONGO_EXPORT int mongo_update( mongo *conn, const char *ns, const bson *cond,
 
     data = &mm->data;
     data = mongo_data_append32( data, &ZERO );
-    data = mongo_data_append( data, ns, strlen( ns ) + 1 );
+    data = mongo_data_append( data, ns, (int)strlen( ns ) + 1 );
     data = mongo_data_append32( data, &flags );
     data = mongo_data_append( data, cond->data, bson_size( cond ) );
     mongo_data_append( data, op->data, bson_size( op ) );
@@ -1016,14 +1016,14 @@ MONGO_EXPORT int mongo_remove( mongo *conn, const char *ns, const bson *cond,
 
     mm = mongo_message_create( 16  /* header */
                               + 4  /* ZERO */
-                              + strlen( ns ) + 1
+                              + (int)strlen( ns ) + 1
                               + 4  /* ZERO */
                               + bson_size( cond )
                               , 0 , 0 , MONGO_OP_DELETE );
 
     data = &mm->data;
     data = mongo_data_append32( data, &ZERO );
-    data = mongo_data_append( data, ns, strlen( ns ) + 1 );
+    data = mongo_data_append( data, ns, (int)strlen( ns ) + 1 );
     data = mongo_data_append32( data, &ZERO );
     mongo_data_append( data, cond->data, bson_size( cond ) );
 
@@ -1144,7 +1144,7 @@ static int mongo_cursor_op_query( mongo_cursor *cursor ) {
 
     mm = mongo_message_create( 16 + /* header */
                                4 + /*  options */
-                               strlen( cursor->ns ) + 1 + /* ns */
+                               (int)strlen( cursor->ns ) + 1 + /* ns */
                                4 + 4 + /* skip,return */
                                bson_size( cursor->query ) +
                                bson_size( cursor->fields ) ,
@@ -1152,7 +1152,7 @@ static int mongo_cursor_op_query( mongo_cursor *cursor ) {
 
     data = &mm->data;
     data = mongo_data_append32( data , &cursor->options );
-    data = mongo_data_append( data , cursor->ns , strlen( cursor->ns ) + 1 );
+    data = mongo_data_append( data , cursor->ns , (int)strlen( cursor->ns ) + 1 );
     data = mongo_data_append32( data , &cursor->skip );
     data = mongo_data_append32( data , &cursor->limit );
     data = mongo_data_append( data , cursor->query->data , bson_size( cursor->query ) );
@@ -1199,7 +1199,7 @@ static int mongo_cursor_get_more( mongo_cursor *cursor ) {
         return MONGO_ERROR;
     } else {
         char *data;
-        int sl = strlen( cursor->ns )+1;
+        int sl = (int)strlen( cursor->ns )+1;
         int limit = 0;
         mongo_message *mm;
 
@@ -1285,7 +1285,7 @@ MONGO_EXPORT int mongo_find_one( mongo *conn, const char *ns, const bson *query,
 MONGO_EXPORT void mongo_cursor_init( mongo_cursor *cursor, mongo *conn, const char *ns ) {
     memset( cursor, 0, sizeof( mongo_cursor ) );
     cursor->conn = conn;
-    cursor->ns = ( const char * )bson_malloc( strlen( ns ) + 1 );
+    cursor->ns = ( const char * )bson_malloc( (int)strlen( ns ) + 1 );
     strncpy( ( char * )cursor->ns, ns, strlen( ns ) + 1 );
     cursor->current.data = NULL;
 }
@@ -1508,7 +1508,7 @@ MONGO_EXPORT int mongo_run_command( mongo *conn, const char *db, const bson *com
     int ret = MONGO_OK;
     bson response = {NULL, 0};
     bson fields;
-    int sl = strlen( db );
+    int sl = (int)strlen( db );
     char *ns = bson_malloc( sl + 5 + 1 ); /* ".$cmd" + nul */
     int res, success = 0;
 
@@ -1665,9 +1665,9 @@ static void mongo_pass_digest( const char *user, const char *pass, char hex_dige
     mongo_md5_byte_t digest[16];
 
     mongo_md5_init( &st );
-    mongo_md5_append( &st, ( const mongo_md5_byte_t * )user, strlen( user ) );
+    mongo_md5_append( &st, ( const mongo_md5_byte_t * )user, (int)strlen( user ) );
     mongo_md5_append( &st, ( const mongo_md5_byte_t * )":mongo:", 7 );
-    mongo_md5_append( &st, ( const mongo_md5_byte_t * )pass, strlen( pass ) );
+    mongo_md5_append( &st, ( const mongo_md5_byte_t * )pass, (int)strlen( pass ) );
     mongo_md5_finish( &st, digest );
     digest2hex( digest, hex_digest );
 }
@@ -1676,7 +1676,7 @@ MONGO_EXPORT int mongo_cmd_add_user( mongo *conn, const char *db, const char *us
     bson user_obj;
     bson pass_obj;
     char hex_digest[33];
-    char *ns = bson_malloc( strlen( db ) + strlen( ".system.users" ) + 1 );
+    char *ns = bson_malloc( (int)(strlen( db ) + strlen( ".system.users" ) + 1) );
     int res;
 
     strcpy( ns, db );
@@ -1724,8 +1724,8 @@ MONGO_EXPORT bson_bool_t mongo_cmd_authenticate( mongo *conn, const char *db, co
     mongo_pass_digest( user, pass, hex_digest );
 
     mongo_md5_init( &st );
-    mongo_md5_append( &st, ( const mongo_md5_byte_t * )nonce, strlen( nonce ) );
-    mongo_md5_append( &st, ( const mongo_md5_byte_t * )user, strlen( user ) );
+    mongo_md5_append( &st, ( const mongo_md5_byte_t * )nonce, (int)strlen( nonce ) );
+    mongo_md5_append( &st, ( const mongo_md5_byte_t * )user, (int)strlen( user ) );
     mongo_md5_append( &st, ( const mongo_md5_byte_t * )hex_digest, 32 );
     mongo_md5_finish( &st, digest );
     digest2hex( digest, hex_digest );
